refactor : move hook into hooks, make index.js to make clean imports
