+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as CalendarNavButton } from "./CalendarNavButton";
+export { default as HomeCalendar } from "./HomeCalendar";
